Move TelegramFacade to view, delete RadioButton.class logic of this class move to EnumSetUtil,refactor pakcaging, delete connection all method to class EnunSet<Subject>(make interface dependency)
