+++ pom.xml
@@ -88,7 +88,7 @@
 					<archive>
 						<manifest>
 							<addClasspath>true</addClasspath>
-							<mainClass>com.softserve.webhookbot.WebhookBotStarter</mainClass>
+							<mainClass>com.softserve.bot.WebhookBotStarter</mainClass>
 						</manifest>
 					</archive>
 				</configuration>
