+++ src/main/java/com/softserve/bot/model/BotMessages.java
@@ -1,0 +1,19 @@
+package com.softserve.bot.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class BotMessages {
+
+    @JsonProperty("start")
+    private String start = "";
+    @JsonProperty("rules")
+    private String rules = "";
+    @JsonProperty("contacts")
+    private String contacts = "";
+}
