+++ src/main/java/com/softserve/bot/service/Filter.java
@@ -1,0 +1,14 @@
+package com.softserve.bot.service;
+
+import com.softserve.bot.service.Specialty;
+import org.springframework.stereotype.Component;
+
+import java.util.EnumSet;
+import java.util.List;
+import java.util.Map;
+@Component
+public class Filter {
+    public Map<String, List<Specialty>> getFiltered(EnumSet enumSet) {
+        return null;
+    }
+}
