+++ src/main/java/com/softserve/bot/service/Specialty.java
@@ -1,0 +1,15 @@
+package com.softserve.bot.service;
+
+public class Specialty {
+    private String code;
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+}
