--- src/main/java/com/softserve/webhookbot/Filter.java
@@ -1,13 +1,0 @@
-package com.softserve.webhookbot;
-
-import org.springframework.stereotype.Component;
-
-import java.util.EnumSet;
-import java.util.List;
-import java.util.Map;
-@Component
-public class Filter {
-    public Map<String, List<Specialty>> getFiltered(EnumSet enumSet) {
-        return null;
-    }
-}
