--- src/main/java/com/softserve/webhookbot/Specialty.java
@@ -1,15 +1,0 @@
-package com.softserve.webhookbot;
-
-public class Specialty {
-    private String code;
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-}
