--- src/main/java/com/softserve/webhookbot/WebhookBotStarter.java
@@ -1,11 +1,0 @@
-package com.softserve.webhookbot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WebhookBotStarter {
-    public static void main(String[] args) {
-        SpringApplication.run(WebhookBotStarter.class, args);
-    }
-}
