--- src/test/java/com/softserve/webhookbot/WebhookBotStarterTests.java
@@ -1,13 +1,0 @@
-package com.softserve.webhookbot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WebhookBotStarterTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
